Keep Cache size to less than 100 entries
Keeping the program open for along time and performing a lost of
queries can fill up the cache to quite considerable levels, up to
several GB. Not everyone has enough RAM for that. With this change the
Cache will be kept to less than 100 Elements, which should keep the
memory footprint to reasonable levels for all reasonable queries.