Add Version into Window Title
To show the user which version they are running. It will require a
ton of diligence from me to keep that string in lockstep with the
version in Cargo.toml and the Git Tag, as well as always updating
both when I release a new Version. But if I manage that it's a really
good addition.