Keep Canvas position upon server reload
Makes sense if you reload the same server just to get a new state, if
you search through the history of a world and if we restore the last
state upon app restart.

Does not make sense when we switch a server. Maybe we can make this
distinction in a future version?