Refactor View::new and View::start into one
This is required to properly set the ui_data to either the saved
settings or the default settings. I do not want to set it to default
settings only to overwrite that a moment later. Though to be honest it
probably doesn't matter much now that I wrote that down. Nevertheless,
no we have one definite place where we set the intial settings, not
two.