Implement Selection import and Export
Add menu items to import and export a list of selections from the
clipboard or from a file. During import it is *not* checked if the
same selection already exists (not yet at least). When exporting all
selections are exported, it is on the user to further filter the list
by editing the generated YAML. May add an option to the individual
selections in the sidebar to export them one by one. The value of this
is questionable though.