Move duplicate Code to TownSelection.refresh()
Both a server reload and certain changes in a selection cause the code
to refresh the complete selection from the backend. This code is now
unified in the TownSelection struct, greatly simplifying the code.

in src/view/mod.rs the part was also moved from the server
reload to after the backend confirmed the server reload with
MessageToView::GotServer. This was required, because before the
invalidated selections requested a refresh from the backend before
the view could send the request for AllTowns and GhostTowns. I found
it less nice to see the all towns selection load in after the user
selections did.