Implement Selection::PartialEq properly
before it was based on the uuid, because I thought I would need an
unchanging constant for unambiguously mapping data from the backend to
one of the displayed selections. But that is not really the case. If
anything else is equal the selection should get the same data anyway.

And because I don't want to store the uuid when exporting selections
I really should not use it for comparisons anyway. So this patch makes
that change finally.