Get rid of dead code in storage.rs
Turns out even with the same content, two databases can still differ
on disk! So the code was not only untouched for a looong time, but
also ultimately useless. Now its gone :)

Rename the one unused MessageToView to better reflect what it will be
used for. We can still remove databases based on age.