Expose more database columns to the user
Island type allows differentiating between islands with and without
farms.

ID is an interesting point to plot how old each town is, or to
highlight a list of towns or islands. Not sure how heavily it will
be used, but we have it so we might as well expose it. Crafting the
selection is not something the users do all the time anyway.