Fix Dependent TownSelection Refreshing
If a selection update, which is referenced by another town selection
via the IN or NOT IN filter, we need to make sure that those
selections also update. This is now implemented.

Also the list of all_selections is now captured after all changes are
made by the user to the sidebar input. This ensures we do not act on
stale information in IN or NOT IN filters anymore. This was previously
a problem, because the backend would receive this stale information
in addition to the up to date town selection with issued a refresh.
The backend would then return bad data, because references could be
broken.