Prevent hidden selections from being processed in the backend
Since they are not visible on the map we may as well not fetch the data
for them. This speeds up initial data loading immensely if the user
uses IN and NOT IN filters, as the subqueries can return a large number
of towns.

This introduces a nasty bug if the filtering is done on the frontend
side though. If the hidden selections never send messages to the
backend, all selections which depend on a hidden selection spam the
backend with update requests. The worst part is that I have no idea
how that bug works. Thankfully, letting the frontend send requests for
hidden selections and simply filtering in the backend gets rid of this
bug.