Suggest meaningful drop down values for "Or" joined selections
While "And" joined selections must filter the suggested list as to
not make a suggestion that would make the whole selection return no
results, "Or" joined selections are the opposite. More constraints
only increase the size of the result set, not decrease it.

TODO: don't suggest values that would duplicate a constraint in both
type (e.g. AllianceName) and value (e.g. AllyA)