Add hidden_id to Selection
In order to keep track of which selection is collapsed. Egui remembers
the state across frames with the persistent id. This was pinned to
the index of the selection, now it is pinned to the new hidden state.
Therefore moving the selection around now keeps its collapsed state!

The hidden_id is also used to improve matching the results returned
from the backend in the front end. Before we relied on matching the
name and constraints, which caused some updates from the backend to
get lost when a lot of requests were queued. Now we can precisely
match the selection, even if the user edited it while the request was
processed in the backend.

Yes, this will lead to temporarily confusing displays, but the
eventual state will be consistent with what the user typed. The
temporarily confusing display will be fixed once we have a priority
queue buffer in the backend that can clear stale requests when new
ones come in.