implement more Constaint.matches methods
the remaining for <=, ==, >= and <> should be done now.

for in selection and not in selection I am not yet sure how to best
implement this. I will probably get a townslist from model/mod.rs
and check if the given town in that list. It would be __much__ more
efficient if I have sets though. So every constraint gets a set of
towns as input and returns a set of towns that match it. Then the
sets from every constraint is joined according to the join mode of the
selection. that way we a inselection/notinselection filter does not
blow up in our face with quadratic runtime.