switch time dependency from time to chrono
time can not get the time in WASM.

But after writing I realized that we don't need the time in WASM...

So it's getting reverted in the next commit. I'll still have the changes in the git history, just in case I do need the time in WASM at some point.