implement sync refresh of drop down values and selection towns
move from message passing model to sync refresh for the drop down
values that are shown in constraints and for the towns that are
attached to each selection. I think I got the refresh logic all right.

The logic for invalidating drop down values for the constraints could
be improved, and is left as a todo in the code.

Refactoring how all_towns/ghost_towns are fetched is next.